Make lib/roken/parse_units.c more robust; fix #522
Use 64-bit types in lib/roken/parse_units.c.

Note that this is an ABI change.  We've already changed the lib/roken
shared object version in preparation for Heimdal 8.0.

+++ lib/roken/parse_bytes.h
@@ -36,6 +36,8 @@
 #ifndef __PARSE_BYTES_H__
 #define __PARSE_BYTES_H__
 
+#include <roken.h>
+
 #ifndef ROKEN_LIB_FUNCTION
 #ifdef _WIN32
 #define ROKEN_LIB_FUNCTION
@@ -46,13 +48,13 @@
 #endif
 #endif
 
-ROKEN_LIB_FUNCTION int ROKEN_LIB_CALL
-parse_bytes (const char *s, const char *def_unit);
+ROKEN_LIB_FUNCTION ssize_t ROKEN_LIB_CALL
+parse_bytes(const char *s, const char *def_unit);
 
 ROKEN_LIB_FUNCTION int ROKEN_LIB_CALL
-unparse_bytes (int t, char *s, size_t len);
+unparse_bytes(ssize_t t, char *s, size_t len);
 
 ROKEN_LIB_FUNCTION int ROKEN_LIB_CALL
-unparse_bytes_short (int t, char *s, size_t len);
+unparse_bytes_short(ssize_t t, char *s, size_t len);
 
 #endif /* __PARSE_BYTES_H__ */
